requirements.txt
# Core dependencies
fastapi>=0.103.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0
fastapi-auth0>=0.5.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
playwright>=1.38.0
selenium>=4.12.0
webdriver-manager>=4.0.0
certifi>=2023.7.22

# Machine learning
tensorflow>=2.13.0
keras>=2.13.1
numpy>=1.24.3

# Image processing
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.2

# AI API integration
anthropic>=0.50.0

# Database
supabase>=1.0.3
python-dotenv>=1.0.0
